% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySignature.R
\name{applySignature}
\alias{applySignature}
\title{Apply a feature signature on a matrix}
\usage{
applySignature(count.mat, genesig00, interceptindex = NA)
}
\arguments{
\item{count.mat}{Matrix with cols=Samples and rows=features}

\item{genesig00}{named numeric vector, Coefficients for a subset of features from count.mat}

\item{interceptindex}{If present, the index of the intercept in genesig00}
}
\value{
response named numeric vector, Holds the response for each sample
}
\description{
Apply a feature signature on a matrix
}
