% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.all.identical.R
\name{check.all.identical}
\alias{check.all.identical}
\title{Check if all elements are identical}
\usage{
check.all.identical(..., func = function(x) {     x })
}
\arguments{
\item{...}{}

\item{func}{The function applied to all elements of "..."}
}
\value{
TRUE if all elements in "..." are after application of func identical.
}
\description{
\enumerate{
  \item Apply func on all elements of "..."
  \item Check if all returned values of func are identical
  \item Return either TRUE or FALSE
}
}
\examples{
check.all.identical(1, 1, "a", 1)
check.all.identical(1, 1, 1, 1)
check.all.identical("a", "a", "a", "a")
}
