% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_missing_names.R
\name{check_missing_names}
\alias{check_missing_names}
\title{Title}
\usage{
check_missing_names(charvec.1, charvec.2, name.1, name.2)
}
\arguments{
\item{charvec.1}{First character vector}

\item{charvec.2}{Second character vector}

\item{name.1}{The (optional) more descriptive name for character vector 1}

\item{name.2}{The (optional) more descriptive name for character vector 2}
}
\value{
NULL, but prints information
}
\description{
Print which elements are in the first but not in the second vector. And the
other way aroung
}
\examples{
check_missing_names(charvec.1 = letters[1:5]
					,charvec.2 = letters[4:6])
check_missing_names(charvec.1 = letters[1:5]
					,charvec.2 = letters[4:6]
					,name.1 = "firstLetters"
					,name.2 = "secondLetters")
}
