% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_no_zero_to_matrix.R
\name{coef_no_zero_to_matrix}
\alias{coef_no_zero_to_matrix}
\title{coef_no_zero result to matrix}
\usage{
coef_no_zero_to_matrix(coefs, do.sort = FALSE,
  do.sort.sortfun = function(coefvec) {     order(abs(coefvec),
  decreasing = TRUE) }, sort.rownames.exceptions = c("(Intercept)",
  "intercept"))
}
\arguments{
\item{coefs}{some coefficients}
}
\value{
A matrix with non-zero coefficients
}
\description{
coef_no_zero result to matrix
}
