% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.nonident.named.R
\name{compare.nonident.named}
\alias{compare.nonident.named}
\title{cbind nonidentical named vectors}
\usage{
compare.nonident.named(...)
}
\arguments{
\item{...}{A (named) concatenation of named vectors or matrices which will then be ordered in a data frame
and the values for the matching names are next to each other in the result}
}
\value{
data.frame where each column is a single given (named) element of ..., each row
a (potentially missing in some elements) name of a coefficient.
}
\description{
cbind into a data frame vectors with nonidentical names.
}
\examples{

vals1 <- 1:10
names(vals1) <- LETTERS[vals1]
vals2 <- 5:12
names(vals2) <- LETTERS[vals2]
compare.nonident.named(tmp = vals1, vals2)
# the following could not work as the rownames are missing!
# compare.nonident.named(tmp=matrix(1:10, ncol=2))
tmpmat <- matrix(1:10, ncol = 2)
rownames(tmpmat) <- LETTERS[1:nrow(tmpmat)]
compare.nonident.named(tmp = tmpmat)
compare.nonident.named(tmpmat)
compare.nonident.named(tmpmat, vals2)
}
