% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_at.R
\name{cut_at}
\alias{cut_at}
\title{Cut at cutvalues}
\usage{
cut_at(values.to.cut, cutvalues, labels)
}
\arguments{
\item{values.to.cut}{What should be cutted}

\item{cutvalues}{where should be cutted}

\item{labels}{optional labels}
}
\value{
A factor
}
\description{
Cut at cutvalues
}
\examples{
cut_at(1:15, 5)
cut_at(1:15, c(5, 7))
cut_at(1:15, c(5, 7, 9))
}
