% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.create_noExistWarn.R
\name{dir.create_noExistWarn}
\alias{dir.create_noExistWarn}
\title{create a directory}
\usage{
dir.create_noExistWarn(dirpath, ...)
}
\arguments{
\item{dirpath}{The directory to create}

\item{...}{Further given to dir.create()}
}
\value{
Nothing, directory is created
}
\description{
Create a directory without the (possible) warning that the directory already
exists. All other warnings are reported
}
\examples{
dir.create_noExistWarn("removeme")
dir.create_noExistWarn("removeme1/sfdg/a")
}
