% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_color_bg.R
\name{facet_color_bg_manual}
\alias{facet_color_bg_manual}
\title{Color a facet title background}
\usage{
facet_color_bg_manual(
  ggplotobj,
  bg.fills.vector = "red",
  remove.axis.labels = FALSE
)
}
\arguments{
\item{ggplotobj}{A ggplot object with facet}

\item{bg.fills.vector}{Colors, either length 1 or with length = # plots for this facet}
}
\value{
A ggplot object
}
\description{
Fill the background of a facet title with color!
}
\examples{

#'
library(ggplot2)
p <- ggplot(mpg, aes(displ, cty)) + geom_point()
p_colfacetted <- p + facet_grid(cols = vars(cyl))
facet_color_bg_manual(p_colfacetted, "yellow")
facet_color_bg_manual(p_colfacetted, "green")
facet_color_bg_manual(p_colfacetted, c("red", "green", "orange", "gray"))
}
