% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_binary_performance.R
\name{get_binary_performance}
\alias{get_binary_performance}
\title{Get binary performance}
\usage{
get_binary_performance(
  prob_positive,
  class,
  measures = c("ACC", "SPEC", "SENS"),
  cutoff = 0.5
)
}
\arguments{
\item{prob_positive}{Probabilities to be in the positive class. Do not essentially have to be 0...1 values but
must be the values predicting the positive class}

\item{measures}{Which measures you want to have, usually measures which are connected to a cutoff.
The measures are returned from ROCit::measureit, see further information there.
'Cutoff' is the actual cutoff value, 'Depth' what proportion of observations fall on or above
the cutoff.}

\item{cutoff}{The cutoff where you want the "measures" from.}

\item{class_positive}{The true class values}
}
\value{

}
\description{
Get performance measures for binary classification problems
}
\examples{

get_performances(
	prob_positive = 1:100,
	class = c(rep(0, 50), rep(1, 50))
)
get_performances(
	prob_positive = 1:100,
	class = c(rep(0, 50), rep(1, 50)),
	cutoff = 50
)
get_performances(
	prob_positive = (1:100)/100,
	class = c(rep(0, 50), rep(1, 50)),
	cutoff = .50
)
}
