% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_color_hue.R
\name{gg_color_hue}
\alias{gg_color_hue}
\title{ggplot color hue}
\usage{
gg_color_hue(n)
}
\arguments{
\item{n}{How many colors you want}
}
\value{
Character vector with the hex-colors of ggplot.
}
\description{
This generates the default ggplot colors
}
\examples{
gg_color_hue(10)
gg_color_hue(2)
}
