% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto_arrange_wrapper.R
\name{ggcyto_arrange_wrapper}
\alias{ggcyto_arrange_wrapper}
\title{Plot multiple colored ggcyto objects}
\usage{
ggcyto_arrange_wrapper(ggcyto.plotlist, facet.coloring.vec,
  do.sort = TRUE, max.columns = 10, facet.coloring.title = "NoTitle",
  global.axis.labels = TRUE)
}
\arguments{
\item{ggcyto.plotlist}{A list of ggcyto objects, received from ggcyto::ggcyto (tested: one sample one element)}

\item{facet.coloring.vec}{A vector with length(ggcyto.plotlist) responsible for coloring the facet background}

\item{do.sort}{Should sorting for facet.coloring.vec be done?}

\item{max.columns}{Maximum amount of columns in a plot}

\item{facet.coloring.title}{Title for the facet color legend}

\item{global.axis.labels}{If TRUE, then all single-plot-axis are removed and one large global axis is added.}
}
\value{
A ggplot object
}
\description{
Plot multiple ggcyto objects, colored by facet.coloring.vec
}
\examples{
ggcyto_arrange_wrapper(ggcyto.plotlist = combinationXX_plots, facet.coloring.vec = as.character(pData(tmp.panel)$Hepatitis))

}
