% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcyto_facet_color_bg.R
\name{ggcyto_facet_color_bg}
\alias{ggcyto_facet_color_bg}
\title{ggcyto facet coloring}
\usage{
ggcyto_facet_color_bg(
  ggcyto.plotlist,
  bg.values,
  numeric.color.palette = colorRampPalette(c("blue", "red")),
  numeric.max.distinct = 50,
  bg.legend.title = "NoTitle",
  remove.axis.labels = FALSE
)
}
\arguments{
\item{ggcyto.plotlist}{A plotlist received when *samplewise* applying ggcyto}

\item{bg.values}{The values to the respective sample}
}
\value{
A ggplot-plotlist with colored facets
}
\description{
color facets of ggcyto output
}
\examples{
library(ggcyto)
data(GvHD)
fs <- GvHD[1:3]
col1 <- "`FSC-H`"
col2 <- "`SSC-H`"
p.list <- flowCore::fsApply(fs, function(x){
	ggcyto(x, aes_string(col1,col2))  +
		geom_hex(bins = 100)
})
ggpubr::ggarrange(plotlist = ggcyto_facet_color_bg(p.list, c("a", "b", "a")))
ggpubr::ggarrange(plotlist = ggcyto_facet_color_bg(p.list, c("a", "b", NA)))
ggpubr::ggarrange(plotlist = ggcyto_facet_color_bg(p.list, c(1, 6, 1)))
}
