% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_from_rpart.R
\name{graph_from_rpart}
\alias{graph_from_rpart}
\title{Generate a network from a rpart-tree}
\usage{
graph_from_rpart(rpart.obj)
}
\arguments{
\item{rpart.obj}{A rpart object from rpart::rpart()}
}
\value{
A data frame with
#'

	- 'from'                # index
	- 'to'                  # index
	- 'node_numbers_from'   # node number in concordance with print.rpart
	- 'node_numbers_to'     # node number in concordance with print.rpart
}
\description{
As a starting point for visualization, this function generates
"from" and "to" indices.
}
\examples{
library(rpart)
fit <- rpart(Kyphosis ~ Age + Number + Start, data = kyphosis)
graph_from_rpart(fit)

# # example for igraph
# routes_igraph <- igraph::graph_from_data_frame(
# 	d = na.omit(generated_edges),
# 	vertices = data.frame('id'=1:length(depth_vec)),
# 	directed = TRUE)

}
