% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictions.R
\name{plot_predictions}
\alias{plot_predictions}
\title{Title}
\usage{
plot_predictions(
  predictions,
  trues,
  names,
  truenames = c(`1` = "infected", `0` = "non.infected"),
  maintitle = "",
  cutoff = 0.5,
  probabilities = TRUE
)
}
\arguments{
\item{predictions}{The prediction values}

\item{trues}{The respective true values}

\item{names}{Names of the samples}

\item{truenames}{E.g. c("1"="infected", "0"="non.infected")
Those names replace the values on the y-axis}

\item{maintitle}{Title}

\item{cutoff}{Cutoff, necessary for horizontal line and performance calculations}

\item{probabilities}{If true, a second y-axis on the right side is added claiming
c("certainly Infected", "certainly not infected")}
}
\value{
A dataframe with AUC, accuracy, sensitivity, specificity and the used cutoff.
}
\description{
Title
}
