% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_probability_pred_true.R
\name{plot_probability_pred_true}
\alias{plot_probability_pred_true}
\title{Title}
\usage{
plot_probability_pred_true(
  predictions,
  trues,
  names,
  truenames = c(`1` = "infected", `0` = "non.infected"),
  maintitle = "",
  cutoff = 0.5,
  probabilities = TRUE,
  sampletype,
  blackwhite = FALSE,
  measures = c("AUC", "ACC", "SPEC", "SENS")
)
}
\arguments{
\item{blackwhite}{}
}
\description{
Title
}
\examples{

set.seed(17)
sampletype <- sample(c("train", "test"), 100, replace = TRUE)
plot_probability_pred_true(
    predictions = 1:100,
    trues = c(rep(0, 50), rep(1, 50)),
    names = paste0("S", 1:100),
    truenames = c("1" = "infected", "0" = "non.infected"),
    maintitle = "CRP",
    cutoff = 60,
    probabilities = FALSE
    # ,sampletype=sampletype
    , blackwhite = TRUE
)

}
