% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ratio_2d_score.R
\name{plot_ratio_2d_score}
\alias{plot_ratio_2d_score}
\title{Plotting an "ratio"-plot}
\usage{
plot_ratio_2d_score(
  range.x,
  range.y,
  n.points.x,
  n.points.y,
  coefs = c(minus = 1, plus = 1),
  titles = c(minus = "", plus = "", total = ""),
  additional.points,
  cutoff.x,
  cutoff.y,
  interpolate = TRUE,
  standardize.plot.color = FALSE,
  additional.points_color.values = NA
)
}
\arguments{
\item{range.x}{Range of x axis}

\item{range.y}{Range of y axis}

\item{n.points.x}{How many points should be generated in the x-axis range.
The final grid will have n.points.x * n.points.y points!}

\item{n.points.y}{How many points should be generated in the y-axis range.
The final grid will have n.points.x * n.points.y points!}

\item{coefs}{The coefficients, length == 3}

\item{titles}{Named character vector c("minus"="", "plus"="", "total"="")
With the titles of the respective plot.}

\item{additional.points}{[Optional] additional points which will be plotted on the score-grid}

\item{cutoff.x}{[Optional] Cutoff of the x-axis, check the plots to see the effect.}

\item{cutoff.y}{[Optional] Cutoff of the y-axis, check the plots to see the effect.}

\item{interpolate}{parameter for geom_raster()}
}
\value{
Plots directly
}
\description{
Plot the x-range score, the y-range score when calculating
"minus" = y - x
"plus" =  y + x
}
\details{
And the total score
}
\examples{
library(ggplot2)
testing_points <- expand.grid(x=(-1:5), y=(-1:9))
plot_ratio_2d_score(range.x = c(-5, 5)
					,n.points.x = 100
					,coefs = c(1, -1))
plot_ratio_2d_score(range.x = range(testing_points$x)
					   ,range.y = range(testing_points$y)
					   ,n.points.x = 100
					   ,coefs = c(1, 1)
					   ,additional.points = testing_points)
plot_ratio_2d_score(range.x = range(testing_points$x)
					   ,range.y = range(testing_points$y)
					   ,n.points.x = 20
					   ,coefs = c(1, 1)
					   ,additional.points = testing_points
					   ,cutoff.x = 2, cutoff.y = 1
)
}
