% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_samplewise.R
\name{plot_samplewise}
\alias{plot_samplewise}
\title{Plot multiple (1+) values for the same names}
\usage{
plot_samplewise(
  values.df,
  samplenames,
  valuename = "Coefficient",
  xname = "Genename",
  hlineintercept = NA,
  yrange = NA,
  connectinglines = FALSE,
  maintitle00 = NA,
  sorting.dfindex = FALSE,
  returnADS = FALSE
)
}
\arguments{
\item{values.df}{Data frame with numerical values (1+ columns, colname is taken as "dataname")}

\item{samplenames}{Names of the samples (N-th samplename --> N-th value in values.df)}

\item{valuename}{How are the values called you plot?}

\item{xname}{x-axis Label name}

\item{hlineintercept}{(named) intercept for the hline}

\item{yrange}{y-range}

\item{connectinglines}{Should per Sample the values be connected?}

\item{maintitle00}{Title of the plot}

\item{sorting.dfindex}{Standard: NA, unsorted. Else: Give index after which values.df-column the x-axis should be sorted}

\item{returnADS}{Return the absolute distance sum between all points for one column, summed up over all columns.}
}
\value{

}
\description{
Plot multiple (1+) values for the same names
}
\examples{
values.df <- data.frame("a"=1:20, "b"=1:20+0.5, 'c'=(20:1)*.5)
names <- paste0('f_', 101:120)
print(plot_samplewise(values.df, names))
print(plot_samplewise(values.df, names, sorting.dfindex=3))
print(plot_samplewise(values.df, names, hlineintercept=5))
print(plot_samplewise(values.df, names, hlineintercept=c("cutoff"=5)))
print(plot_samplewise(values.df, names, hlineintercept=c("cutoff"=5)))
print(plot_samplewise(values.df, names, connectinglines=TRUE))
print(plot_samplewise(values.df, names, returnADS=TRUE))
}
