% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predplot_multiclass.R
\name{predplot_multiclass}
\alias{predplot_multiclass}
\title{Prediction matrix of a multiclass problem}
\usage{
predplot_multiclass(pred_matrix, true_values, sort_by_true = TRUE)
}
\arguments{
\item{pred_matrix}{The predicted probabilities for each class (=columns) of each sample (=row)}

\item{true_values}{Either the labels from the columns of pred_matrix, or the indices for the respective columns}

\item{sort_by_true}{If TRUE, then the rows are sorted by the true values.}
}
\description{
Prediction matrix of a multiclass problem
}
\examples{
n_samples <- 200
classes <- c("A", "B", "C", "D")
ex_trues <- sample(classes, n_samples, replace = TRUE)

example_predictions <- lapply(
    ex_trues,
    function(x) {
        multiclass_probability_simulation(length(classes))
    }
)
example_predictions <- do.call(rbind, example_predictions)
colnames(example_predictions) <- classes


pdf("removeme1.pdf")
print(predplot_multiclass(
    pred_matrix = example_predictions,
    true_values = ex_trues
))
for (complete_letter in classes) {
    print(predplot_multiclass(
        pred_matrix = example_predictions,
        true_values = rep(complete_letter, nrow(example_predictions))
    ))
}
dev.off()

}
