% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_as_python.R
\name{print_as_python}
\alias{print_as_python}
\title{print_as_python}
\usage{
print_as_python(x)
}
\arguments{
\item{x}{The matrix to print}
}
\value{
prints only
}
\description{
Print a 2d matrix python-like
}
\examples{
print_as_python(
	matrix(c(-1.2067,  0.6976,
			 0.1576,  0.6925,
			 1.0208, -0.6271), ncol=2, byrow=TRUE)
	)
}
