% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_first_PC.R
\name{remove_first_PC}
\alias{remove_first_PC}
\title{Remove first PC}
\usage{
remove_first_PC(X)
}
\arguments{
\item{X}{Rows are features
Columns are samples}
}
\value{
A matrix with same dimension as X but where each sample got the first principal component of a pca on all data removed.
}
\description{
Remove first principal component from a dataset X.
This means:
1. Do a PCA
2. Apply the PCA on the data --> some value for the first PC for each sample
3. Do this function, receive X'.
4. Apply PCA from 1., then all values of the first principal component must be approx. zero.
}
\examples{
data("mtcars")
tmp <- t(mtcars[, c(1:7, 10, 11)])
remove_first_PC(tmp)
}
