% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_get_element.R
\name{simplify_results}
\alias{simplify_results}
\title{simplify matrix of lists}
\usage{
simplify_results(interpreted.df, get.names = FALSE)
}
\arguments{
\item{interpreted.df}{Matrix from interpret_unlist_get_element. The last column holds a list of the searched values.}
}
\value{
A simplified version of interpreted.df
}
\description{
Matrix from interpret_unlist_get_element. The last column holds a list of the searched values.
Often, those values can be simplified!
}
\details{
Each row is gone through and duplicated the number of elements in the last column.
(Remark 2019-12-16: could be problematic with arrays)
}
