% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCoef_ratioCoef.R
\name{singleCoef_ratioCoef}
\alias{singleCoef_ratioCoef}
\title{Get single coefficients from ratio signature}
\usage{
singleCoef_ratioCoef(named.ratio_coefmat, splitchar = "_/_")
}
\arguments{
\item{named.ratio_coefmat}{A ratio coefficient matrix with rownames}

\item{splitchar}{How should the ratio-coefficient rownames be splitted}
}
\value{
A matrix with the single coefficients as rownames and the coefficients as values
in the first column
}
\description{
Get single coefficients from ratio signature
}
\examples{
ratio.signature <- c("longname.b_/_longname.a"=1, "longname.c_/_longname.b"=-1)
single.coef_signature <- singleCoef_ratioCoef(ratio.signature)
print(single.coef_signature)
}
