% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unimplemented_s_finding.R
\name{unimplemented_s_finding}
\alias{unimplemented_s_finding}
\title{lambda finding}
\usage{
unimplemented_s_finding(fit, s = "lambda.1se",
  fit.lambda.name = "lambda", fit.lambda.error.name = "cve",
  fit.lambda.errorSD.name = "cvse", ...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
If "s" is not the usual lambda-name, I have to search for it in another way.
Here's how.
}
