% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unlist_get_element.R
\name{unlist_get_element}
\alias{unlist_get_element}
\title{Recursive unlisting for a specific element}
\usage{
unlist_get_element(
  mylist,
  which.element,
  simplify = TRUE,
  get.names = FALSE,
  max.depth = Inf
)
}
\arguments{
\item{mylist}{Your listlistlist... you want to delist}

\item{which.element}{How is the element you want called? "coefNoZero", "cutoff", etc.}

\item{simplify}{[default: TRUE] Try to make usefull stuff out of the last column of the interpreted dataframe}

\item{get.names}{[default: FALSE] Should names be extracted?}
}
\value{
A data.frame, if not simplified with a list as last-column elements per row
}
\description{
Retrieve a value (vector and matrix tested) from a list of list of list of ...
and return a dataframe.
}
\examples{

a <- yaml::yaml.load_file("R/a.yaml")
b <- a
b$trainScaled <- b$train.res...p
a_small <- a
a_small$train.res...p <- a_small$train.res...p[19:20]
tmp <- unlist_get_element(a_small, "cutoff")
tmp <- unlist_get_element(a, "cutoff")
tmp <- unlist_get_element(b, "cutoff")


b_modified <- b
b_modified$train.res...p$CVstep1$cutoff <- list("a", "b", "c")
unlist_get_element(b_modified, "cutoff", simplify = TRUE)

# first list here is deeper than
a_different_depths <- a_small
a_different_depths$train.res...p$CVstep1$asdflkjh$cutoff <- "sads"
a_different_depths$train.res...p$CVstep1$cutoff <- NULL
print(unlist_get_element(a_different_depths, "cutoff"))
# second list here is deeper than
a_different_depths <- a_small
a_different_depths$train.res...p$CVstep2$asdflkjh$cutoff <- "sads"
a_different_depths$train.res...p$CVstep2$cutoff <- NULL
print(unlist_get_element(a_different_depths, "cutoff"))
print(unlist_get_element(a_different_depths, "cutoff", simplify = FALSE))


# one element is a 1-col matrix
a_different_depths <- a_small
a_different_depths$train.res...p$CVstep2$asdflkjh$cutoff <- matrix(1:4, ncol=1)
a_different_depths$train.res...p$CVstep2$cutoff <- NULL
print(unlist_get_element(a_different_depths, "cutoff"))

# one element is a 1-row matrix # error
a_different_depths <- a_small
a_different_depths$train.res...p$CVstep2$asdflkjh$cutoff <- matrix(1:4, nrow=1)
a_different_depths$train.res...p$CVstep2$cutoff <- NULL
print(unlist_get_element(a_different_depths, "cutoff"))


# one element is a 1-row matrix # error
a_different_depths <- a_small
a_different_depths$train.res...p$CVstep2$asdflkjh$cutoff <- matrix(1:4, ncol=1)
rownames(a_different_depths$train.res...p$CVstep2$asdflkjh$cutoff) <- letters[1:4]
a_different_depths$train.res...p$CVstep2$cutoff <- NULL
print(unlist_get_element(a_different_depths, "cutoff"))
}
