% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.allcomb.R
\name{vector.allcomb}
\alias{vector.allcomb}
\title{All combinations of vector}
\usage{
vector.allcomb(named.num.vector, combfun = "/", combfun.name, ...)
}
\arguments{
\item{named.num.vector}{A (named) numeric vector}

\item{combfun}{How it should be combined, is overgiven to \code{outer(FUN=combfun)}}

\item{...}{optional further arguments passed to FUN in the outer call: \code{outer(X, Y, FUN=combfun, ...)}}
}
\value{
Numeric which has as names what is combined with what by which combfun:

valuename1_<combfun>_valuename2
}
\description{
Generate all possible combinations of a named vector with a specific combination function.
}
\examples{
vector.allcomb(1:10, combfun = function(x, y)x**2, combfun.name = "^2")
vector.allcomb(1:10, combfun = "*")
myvec <- 1:10
names(myvec) <- paste0("gene", myvec)
vector.allcomb(1:10, combfun = "-")
vector.allcomb(1:10, combfun = "/")
complex_function <- function(x, y){
	x**y
}
vector.allcomb(1:10, combfun = complex_function, combfun.name = "complex")
exprs(miR_Ros.set)[1:3, 1]
a <- outer(exprs(miR_Ros.set)[1:3, 1], exprs(miR_Ros.set)[1:3, 1], "/")
a
## so I see, it is x-axis-gene / y-axis-gene
a[upper.tri(a, diag=TRUE)] <- NA
b <- na.omit(reshape2::melt(a))
b
## so it is Var1/Var2

}
