% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorize_named.R
\name{vectorize_named}
\alias{vectorize_named}
\title{Convert a matrix to a vector}
\usage{
vectorize_named(named_matrix)
}
\arguments{
\item{named_matrix}{Matrix with one column which gets transformed into a named vector}
}
\value{
named vector
}
\description{
Convert a named matrix with one column to a named vector where the rownames are the names.
}
\examples{
mymat <- matrix(1:6, nrow = 6)
vectorize_named(mymat)

rownames(mymat) <- paste0("f_", 1:6)
vectorize_named(mymat)

mymat <- matrix(1:6, nrow = 3)
# This here should fail because there is more than 1 column in the matrix mymat!
vectorize_named(mymat)

}
