% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_fcs.R
\name{write_fcs}
\alias{write_fcs}
\title{Export a matrix to an .fcs file}
\usage{
write_fcs(cell_matrix, filepath, max_val = 1e+07, min_val = -1000)
}
\arguments{
\item{cell_matrix}{Matrix with cols=Samples and rows=features}

\item{filepath}{Where the .fcs gets saved}

\item{max_val}{The fake cell with maximum values which gets added and after flowFrame creation removed}

\item{min_val}{The fake cell with miminum values which gets added and after flowFrame creation removed}
}
\description{
It adds (and removes) two cells with constant maximum and minimum values to have nice visual
visual ranges for Kaluza
}
