% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applySignature.R
\name{applySignature}
\alias{applySignature}
\title{Apply a feature signature on a matrix}
\usage{
applySignature(
  count.mat,
  genesig,
  interceptindex = NA,
  genesig.matrix.colN = 1,
  genesig.data.frame.cols_names = 1,
  genesig.data.frame.cols_coefs = 2
)
}
\arguments{
\item{count.mat}{Matrix with cols=Samples and rows=features}

\item{genesig}{named numeric vector, Coefficients for a subset of features from count.mat
If a matrix is given,}

\item{interceptindex}{If present, the index of the intercept in genesig}

\item{genesig.matrix.colN}{If the given genesig is a matrix, which column should be taken}

\item{genesig.data.frame.cols_names}{If the given genesig is a data frame, which column is the signature-feature-names}

\item{genesig.data.frame.cols_coefs}{If the given genesig is a data frame, which column holds the signature-coefficients}
}
\value{
response named numeric vector, Holds the response for each sample
}
\description{
Apply a feature signature on a matrix
}
\examples{
datamat <- matrix(1:20, nrow=5)
colnames(datamat) <- paste0("S", 1:ncol(datamat))
rownames(datamat) <- paste0("F", 1:nrow(datamat))

sig1 <- rep(2, nrow(datamat))
names(sig1) <- rownames(datamat)
applySignature(count.mat = datamat
			   ,genesig = sig1)


sig2 <- as.matrix(sig1)
applySignature(count.mat = datamat
			   ,genesig = sig2)

sig3 <- cbind(sig1, -1*sig1, 2*sig1)
applySignature(count.mat = datamat
			   ,genesig = sig3
			   ,genesig.matrix.colN = 1)
applySignature(count.mat = datamat
			   ,genesig = sig3
			   ,genesig.matrix.colN = 2)

sig4 <- as.data.frame(sig3)
applySignature(count.mat = datamat
			   ,genesig = sig4
			   ,genesig.data.frame.cols_names = 0
			   ,genesig.data.frame.cols_coefs = 1)
applySignature(count.mat = datamat
			   ,genesig = sig4
			   ,genesig.data.frame.cols_names = 0
			   ,genesig.data.frame.cols_coefs = 2)
sig5 <- sig4
sig5$names <- rownames(sig5)
rownames(sig5) <- NULL
applySignature(count.mat = datamat
			   ,genesig = sig5
			   ,genesig.data.frame.cols_names = 4
			   ,genesig.data.frame.cols_coefs = 2)
applySignature(count.mat = datamat
			   ,genesig = sig5
			   ,genesig.data.frame.cols_names = "names"
			   ,genesig.data.frame.cols_coefs = 2)
applySignature(count.mat = datamat
			   ,genesig = sig5
			   ,genesig.data.frame.cols_names = 4
			   ,genesig.data.frame.cols_coefs = "V2")
}
