% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_no_zero_to_matrix.R
\name{coef_no_zero_to_matrix}
\alias{coef_no_zero_to_matrix}
\title{coef_no_zero result to matrix}
\usage{
coef_no_zero_to_matrix(coefs)
}
\arguments{
\item{coefs}{}
}
\value{

}
\description{
coef_no_zero result to matrix
}
