% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singleCoef_ratioCoef.R
\name{singleCoef_ratioCoef}
\alias{singleCoef_ratioCoef}
\title{Title}
\usage{
singleCoef_ratioCoef(named.ratio_coefmat, splitchar = "_/_")
}
\arguments{
\item{splitchar}{}
}
\description{
Title
}
\examples{
ratio.signature <- c("longname.b_/_longname.a"=1, "longname.c_/_longname.b"=-1)
single.coef_signature <- singleCoef_ratioCoef(ratio.signature)
print(single.coef_signature)
}
