% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector.allcomb.R
\name{vector.allcomb}
\alias{vector.allcomb}
\title{All combinations of vector}
\usage{
vector.allcomb(named.num.vector, combfun = "/", combfun.name, ...)
}
\arguments{
\item{named.num.vector}{A (named) numeric vector}

\item{combfun}{How it should be combined, is overgiven to \code{outer(FUN=combfun)}}

\item{combfun.name}{If the combfun is an (anonymous) function, you have to specify its name in combfun.name}

\item{...}{optional further arguments passed to FUN in the outer call: \code{outer(X, Y, FUN=combfun, ...)}}
}
\value{
Numeric which has as names what is combined with what by which combfun:

valuename1_<combfun>_valuename2
}
\description{
Generate all possible combinations of a named vector with a specific combination function.
}
\examples{
vector.allcomb(1:10, combfun = function(x, y)x**2, combfun.name = "^2")
vector.allcomb(1:10, combfun = "*")
myvec <- 1:10
names(myvec) <- paste0("gene", myvec)
vector.allcomb(1:10, combfun = "-")
vector.allcomb(1:10, combfun = "/")
complex_function <- function(x, y){
	x**y
}
vector.allcomb(1:10, combfun = complex_function, combfun.name = "complex")
}
